refactor(engine): ProjectStageProcessor + B6
Extract GameEngine.advanceProjectStages verbatim into
shared/engine/processors/ProjectStageProcessor.ts (ctx rebinding only; hardcoded
week gates stay hardcoded). The processor calls TourProcessor directly with the
WeekContext for mid-tour city revenue — identical behavior to the old inline
this.processUnifiedTourRevenue delegate (which was just a
new TourProcessor().processUnifiedTourRevenue(ctx, ...) wrapper), with no engine
round-trip. game-engine.ts keeps a same-signature advanceProjectStages delegate.

B6 (decision D3, SANCTIONED behavior change): delete the no-op
processNewlyRecordedProjects / processProjectSongRecording pass (~95 lines) and
its advanceWeek pipeline call. Songs are born isRecorded:true in
SongGenerationProcessor.generateSong, so the pass's !isRecorded filter always
matched zero songs and its "recording completed - ready for release"
notification never fired. That notification now emits from
generateWeeklyProjectSongs when a project's last song is generated — one per
song in the completing batch, matching the dead pass's
{ type: 'unlock', description, amount: 0 } shape. recordedAt is now stamped at
song creation (the dead pass was the only writer; wall-clock metadata, stripped
by the golden-master normalizer).

Golden master regenerated: the ONLY diffs are in recording-week (the Single
completes both songs in one week) — two new unlock notifications and recordedAt
null -> <timestamp> on the two generated songs. No other scenario changed; no
RNG-order change (the stage machine makes no RNG draws). Added
project-stage-recording.test.ts pinning the B6 notification (fires on
completion, silent mid-recording), recordedAt stamping, and the stage/rate
helpers.

game-engine.ts 4001 -> 3751 lines. ProjectStageProcessor 222 lines.
